pyproject.toml
[project]
name = "factor-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.8.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]
